feat: 개봉예정영화 UI 완료
